+++ __init__.py
@@ -1,5 +1,5 @@
 """
 Oktopios — Un langage de programmation moderne et expressif 🐙
 """
-__version__ = "0.1.0"
+__version__ = "0.1.3"
 __author__ = "Mouanwiya Ali Soule"